+++ docs/src/tutorials/sdm-training.jl
@@ -77,7 +77,7 @@ current_figure() #hide
 # of labels, we give a vector of layers (features), and the two layers for
 # presences and absences:
 
-sdm = SDM(ZScore, Logistic, L, presencelayer, bgpoints)
+sdm = SDM(RawData, Logistic, L, presencelayer, bgpoints)
 hyperparameters!(classifier(sdm), :η, 1e-4);
 hyperparameters!(classifier(sdm), :interactions, :all);
 hyperparameters!(classifier(sdm), :epochs, 10_000);
@@ -85,6 +85,12 @@ hyperparameters!(classifier(sdm), :epochs, 10_000);
 # We will now train the model on all the training data.
 
 train!(sdm)
+
+# In order to make sure we have a robust threshold for the model, we will also
+# re-estimate it on cross-validation samples. This is a slightly more reliable
+# version than the thresholding done as part of `train!`.
+
+threshold!(sdm)
 
 # ## Making the first prediction